74 | Extract form validation into stimulus controller for reusability
