+++ app/javascript/controllers/index.js
@@ -13,6 +13,9 @@ application.register("evaluation-criteria", EvaluationCriteriaController);
 import EvaluationFormController from "./evaluation_form_controller";
 application.register("evaluation-form", EvaluationFormController);
 
+import FormValidationController from "./form_validation_controller";
+application.register("form-validation", FormValidationController);
+
 import HotdogController from "./hotdog_controller";
 application.register("hotdog", HotdogController);
